requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
typing==3.7.4.1
gunicorn==20.0.4
django-heroku==0.3.1